refactor: use folders as domains
