+++ lib/batches/write-batch.utils.ts
