+++ lib/collections/firestore-collection.interface.ts
