--- lib/decorators/index.ts
@@ -1,2 +1,0 @@
-export * from './collection.decorator';
-export * from './inject-repository.decorator';
