--- lib/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './firestore-document.dto';
