+++ lib/index.ts
@@ -1,6 +1,3 @@
 export * from './firestore.module';
-export * from './decorators';
 export * from './errors';
-export * from './interfaces';
-export * from './dto';
-export * from './repository';
+export * from './repositories';
