--- lib/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './firestore-collection.interface';
-export * from './firestore-module-options.interface';
