+++ lib/options/firestore-module-options.interface.ts
