+++ lib/queries/firestore-document.dto.ts
