+++ lib/repositories/batch-repository.provider.ts
@@ -1,5 +1,3 @@
-import { Transaction } from '../transactions/transaction.provider';
-import { FirestoreDocument } from '../dto';
 import { InvalidArgumentError } from '../errors/invalid-argument.error';
 import {
   DocumentReference,
@@ -7,25 +5,28 @@ import {
   UpdateData,
 } from '@google-cloud/firestore';
 import { BaseRepository } from './base-repository.provider';
-import { CollectionMetadata, FirestoreModuleCoreOptions } from '../interfaces';
+import { WriteBatch } from '../batches/batch.provider';
+import { FirestoreModuleCoreOptions } from '../options/firestore-module-options.interface';
+import { CollectionMetadata } from '../collections/firestore-collection.interface';
+import { FirestoreDocument } from '../queries/firestore-document.dto';
 
-export class TransactionalRepository<
+export class WriteBatchRepository<
   T extends FirestoreDocument,
 > extends BaseRepository<T> {
   constructor(
     firestore: Firestore,
     firestoreModuleOptions: FirestoreModuleCoreOptions,
     collectionOptions: CollectionMetadata<T>,
-    private readonly tx: Transaction,
+    private readonly batch: WriteBatch,
   ) {
     super(firestore, firestoreModuleOptions, collectionOptions);
   }
 
-  async create(document: T): Promise<T> {
+  create(document: T): T {
     const { id, ...object } = document;
 
     const docRef = id ? this.collectionRef.doc(id) : this.collectionRef.doc();
-    this.tx._create(docRef, object as T);
+    this.batch._create(docRef, object as T);
 
     return {
       ...document,
@@ -33,7 +34,7 @@ export class TransactionalRepository<
     };
   }
 
-  async update(document: Partial<T>): Promise<Partial<T>> {
+  update(document: Partial<T>): Partial<T> {
     const { id, ...object } = document;
 
     if (!id) {
@@ -41,49 +42,27 @@ export class TransactionalRepository<
     }
 
     const docRef = this.collectionRef.doc(id);
-    this.tx._update(docRef, object as UpdateData<T>);
+    this.batch._update(docRef, object as UpdateData<T>);
 
     return {
       ...document,
     };
   }
 
-  async delete(id: string): Promise<void> {
+  delete(id: string): void {
     const docRef: DocumentReference<T> = this.collectionRef.doc(id);
 
     if (this.collectionOptions.softDelete) {
-      this.tx._update(docRef, { deleteTime: new Date() } as UpdateData<T>);
+      this.batch._update(docRef, { deleteTime: new Date() } as UpdateData<T>);
     } else if (
       !this.firestoreModuleOptions.softDelete &&
       this.collectionOptions.softDelete
     ) {
-      this.tx._update(docRef, {
+      this.batch._update(docRef, {
         deleteTime: new Date(),
       } as UpdateData<T>);
     } else {
-      this.tx._delete(docRef);
+      this.batch._delete(docRef);
     }
-  }
-
-  async findById(id: string): Promise<T | null> {
-    const docRef: DocumentReference<T> = this.collectionRef.doc(id);
-
-    const doc = await this.tx._get(docRef);
-    if (!doc.exists) {
-      return null;
-    }
-
-    const docData = doc.data() as FirestoreDocument;
-    if (docData.deleteTime) {
-      return null;
-    }
-
-    return {
-      ...docData,
-      id: doc.id,
-      createTime: doc.createTime?.toDate(),
-      updateTime: doc.updateTime?.toDate(),
-      readTime: doc.readTime?.toDate(),
-    } as T;
   }
 }
