+++ lib/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './firestore-repository.provider';
